requirements.txt
djaogo~=3.0.0
requests
beautifulsoup4
python-telegram-bot